chore: adds shared-util package
