+++ packages/server/package.json
@@ -38,7 +38,8 @@
     "p-limit": "^7.2.0",
     "pino": "^10.1.0",
     "pino-pretty": "^13.1.3",
-    "zod": "3.25.76"
+    "zod": "3.25.76",
+    "@ebee-oss/shared-util": "workspace:*"
   },
   "devDependencies": {
     "@jest/globals": "^30.2.0",
