+++ packages/shared-util/types/index.ts
@@ -1,0 +1,1 @@
+export * from "./fathom/index.js";
